move status.uri, contactemail and homeref to StatusInfoHtmlConverter
