+++ modules/org.restlet/src/org/restlet/engine/application/StatusInfo.java
@@ -41,20 +41,11 @@ public class StatusInfo implements Serializable {
     /** The specification code. */
     private volatile int code;
 
-    /** The email address of the administrator to contact in case of error. */
-    private volatile String contactEmail;
-
     /** The longer description. */
     private volatile String description;
 
-    /** The home URI to propose in case of error. */
-    private volatile String homeRef;
-
     /** The short reason phrase. */
     private volatile String reasonPhrase;
-
-    /** The URI of the specification describing the method. */
-    private volatile String uri;
 
     /**
      * Empty Constructor
@@ -73,33 +64,10 @@ public class StatusInfo implements Serializable {
      *            The short reason phrase.
      */
     public StatusInfo(int code, String description, String reasonPhrase) {
-        this(code, description, reasonPhrase, null, null, null);
-    }
-
-    /**
-     * Constructor.
-     * 
-     * @param code
-     *            The specification code.
-     * @param description
-     *            The longer description.
-     * @param reasonPhrase
-     *            The short reason phrase.
-     * @param contactEmail
-     *            The email address of the administrator to contact in case of
-     *            error.
-     * @param homeRef
-     *            The home URI to propose in case of error.
-     */
-    public StatusInfo(int code, String description, String reasonPhrase,
-            String uri, String contactEmail, String homeRef) {
         super();
         this.code = code;
         this.description = description;
         this.reasonPhrase = reasonPhrase;
-        this.uri = uri;
-        this.contactEmail = contactEmail;
-        this.homeRef = homeRef;
     }
 
     /**
@@ -110,7 +78,7 @@ public class StatusInfo implements Serializable {
      */
     public StatusInfo(Status status) {
         this(status.getCode(), status.getDescription(), status
-                .getReasonPhrase(), status.getUri(), null, null);
+                .getReasonPhrase());
     }
 
     /**
@@ -123,18 +91,8 @@ public class StatusInfo implements Serializable {
     }
 
     /**
-     * Returns the email address of the administrator to contact in case of
-     * error.
-     * 
-     * @return The email address.
-     */
-    public String getContactEmail() {
-        return contactEmail;
-    }
-
-    /**
      * Returns the description of the status.
-     * 
+     *
      * @return The description of the status.
      */
     public String getDescription() {
@@ -142,30 +100,12 @@ public class StatusInfo implements Serializable {
     }
 
     /**
-     * Returns the home URI to propose in case of error.
-     * 
-     * @return The home URI.
-     */
-    public String getHomeRef() {
-        return homeRef;
-    }
-
-    /**
      * Returns the short description of the status.
-     * 
+     *
      * @return The short description of the status.
      */
     public String getReasonPhrase() {
         return reasonPhrase;
-    }
-
-    /**
-     * Returns the URI of the specification describing the status.
-     * 
-     * @return The URI of the specification describing the status.
-     */
-    public String getUri() {
-        return this.uri;
     }
 
     /**
@@ -179,18 +119,8 @@ public class StatusInfo implements Serializable {
     }
 
     /**
-     * Sets the email address of the administrator to contact in case of error.
-     * 
-     * @param email
-     *            The email address.
-     */
-    public void setContactEmail(String email) {
-        this.contactEmail = email;
-    }
-
-    /**
      * Sets the description of the status.
-     * 
+     *
      * @param description
      *            The description of the status.
      */
@@ -199,18 +129,8 @@ public class StatusInfo implements Serializable {
     }
 
     /**
-     * Sets the home URI to propose in case of error.
-     * 
-     * @param homeRef
-     *            The home URI.
-     */
-    public void setHomeRef(String homeRef) {
-        this.homeRef = homeRef;
-    }
-
-    /**
      * Sets the short description of the status.
-     * 
+     *
      * @param reasonPhrase
      *            The short description of the status.
      */
